+++ HISTORY.rst
@@ -108,6 +108,7 @@
     preview errors with a dedicated ``AppImageErrorImage`` instance.
 
 - Support sorting documents by creation date in all view.
+- Add ``AppImageErrorImage`` indicator for when a document file has no pages.
 
 4.6.2 (2024-03-04)
 ==================